add color manager utils tests
