+++ napari/layers/utils/_tests/test_layer_utils.py
@@ -3,7 +3,6 @@ import pandas as pd
 import pytest
 from dask import array as da
 
-from napari.layers.utils.color_manager_utils import guess_continuous
 from napari.layers.utils.layer_utils import (
     calc_data_range,
     dataframe_to_properties,
@@ -94,14 +93,3 @@ def test_dataframe_to_properties():
     properties_df = pd.DataFrame(properties)
     converted_properties, _ = dataframe_to_properties(properties_df)
     np.testing.assert_equal(converted_properties, properties)
-
-
-def test_guess_continuous():
-    continuous_annotation = np.array([1, 2, 3], dtype=np.float32)
-    assert guess_continuous(continuous_annotation)
-
-    categorical_annotation_1 = np.array([True, False], dtype=bool)
-    assert not guess_continuous(categorical_annotation_1)
-
-    categorical_annotation_2 = np.array([1, 2, 3], dtype=np.int)
-    assert not guess_continuous(categorical_annotation_2)
